Add a framework for testing adb sideload/recovery OTA.
This test aims to be as simple and easy to execute as possible. It does
not require server infrastructure or large build artifacts. Instead, it
uses a genrule to create a very small (32KB) OTA package. It's
technically a full package, but it's a partial OTA, and simply adds a
partition containing random test data.

This allows us to test the OTA recovery process in seconds, instead of
minutes or hours. It also skirts around having to deal with vbmeta/AVB,
since it does not change any verified partitions.

The initial test in the suite applies the test package in recovery, then
verifies the slot switch and partition data. Future tests will expand on
this.

Bug: 352539864
Test: RecoveryOtaTest
Change-Id: Ibce49b58729eec25c25fb9b84a520a3adfd3efa7