Fix race condition in RecoveryOtaTest.
The device can reboot in between checking !inRecovery() and running "adb
shell getprop". Normally getProperty() handles this, but we were
bypassing it to avoid the property cache.

Instead, just explicitly invalidate the property cache.

This is to address flaky ATP results on Pixel 6.

Also add an OWNERS file to include OTA folks.

Bug: 352539864
Test: RecoveryOtaTest
Change-Id: I879f02bed081fea2e7531e1e84e549f6a2cab0f1