Fix merges getting stuck in sideload tests.
Sometimes we pause checkpointing to test whether merges resume/cancel
properly in recovery. However, we never properly resume checkpointing,
because resetting the persist property does not actually re-run the
necessary triggers.

As a fix, detect when dont_commit_checkpoint=1 and forcefully resume
checkpointing. There is technically still an issue here if we commit the
checkpoint before sys.boot_complete, but it doesn't seem to happen in
practice.

Bug: 438092240
Test: atest RecoveryOtaTest on Pixel 6/7
Change-Id: Id45e9255d93630dc86ca1d855e34f7da85c5619c